quiz_service/requirements.txt
aioredis==2.0.1
appnope==0.1.4
asgiref==3.10.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==24.4.2
Automat==25.4.16
backports.asyncio.runner==1.2.0
cffi==2.0.0
channels==4.3.1
comm==0.2.3
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
debugpy==1.8.17
decorator==5.2.1
Django==4.2.25
django-redis==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
exceptiongroup==1.3.0
executing==2.2.1
fakeredis==2.32.1
hyperlink==21.0.0
idna==3.11
importlib_metadata==8.7.0
incremental==24.7.2
inflection==0.5.1
iniconfig==2.1.0
ipykernel==6.30.1
ipython==8.18.1
jedi==0.19.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.5
pexpect==4.9.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.1.0
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.3.0
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-django==4.11.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
pyzmq==27.1.0
redis==7.0.1
referencing==0.36.2
rpds-py==0.27.1
service-identity==24.2.0
six==1.17.0
sortedcontainers==2.4.0
sqlparse==0.5.3
stack-data==0.6.3
tomli==2.3.0
tornado==6.5.2
traitlets==5.14.3
Twisted==25.5.0
txaio==23.6.1
typing_extensions==4.15.0
uritemplate==4.2.0
wcwidth==0.2.14
zipp==3.23.0
zope.interface==8.0.1
